+++ spec/youku/v2/searches_spec.rb
@@ -34,6 +34,28 @@ describe Youku::V2::Searches do
 
       it_should_behave_like 'a base Youku API V2 requests'
     end
+
+    describe '#by_tag' do
+      let(:url) { 'https://openapi.youku.com/v2/searches/video/by_tag.json' }
+      let(:query) { {
+        client_id: client.client_id,
+        tag:      2014,
+        category: '',
+        period:   'today',
+        orderby:  'relevance',
+        userid:   '',
+        page:     1,
+        count:    20
+      } }
+
+      before do
+        stub_request(:get, url).with(query: query).to_return(status: 200)
+      end
+
+      subject { video.by_tag(tag: 2014) }
+
+      it_should_behave_like 'a base Youku API V2 requests'
+    end
   end
 
 end